jsp viré de add/mod book + correction pb isbn int->long

+++ src/beans/IsbnComp.java
@@ -6,7 +6,7 @@ public class IsbnComp implements Comparator<Book> {
 
 	@Override
 	public int compare(Book b1, Book b2) {
-		return Integer.toString(b1.getIsbn()).compareTo(Integer.toString(b2.getIsbn()));
+		return Long.toString(b1.getIsbn()).compareTo(Long.toString(b2.getIsbn()));
 	}
 
 }
